refactor: 활성화/종료 변경 훅으로 분리
